+++ deps/src/onemkl.h
@@ -187,7 +187,7 @@ void onemklSsyr2k(syclQueue_t device_queue, onemklUplo upper_lower, onemklTransp
                 const float *b, int64_t ldb, float beta, float *c, int64_t ldc);
 void onemklDsyr2k(syclQueue_t device_queue, onemklUplo upper_lower, onemklTranspose trans,
                 int64_t n, int64_t k, double alpha, const double *a, int64_t lda,
-                const double *b, int64_t ldb, float beta, double *c, int64_t ldc);
+                const double *b, int64_t ldb, double beta, double *c, int64_t ldc);
 void onemklCsyr2k(syclQueue_t device_queue, onemklUplo upper_lower, onemklTranspose trans,
                 int64_t n, int64_t k, float _Complex alpha, const float _Complex *a,
                 int64_t lda, const float _Complex *b, int64_t ldb, float _Complex beta,
@@ -195,7 +195,7 @@ void onemklCsyr2k(syclQueue_t device_queue, onemklUplo upper_lower, onemklTransp
 void onemklZsyr2k(syclQueue_t device_queue, onemklUplo upper_lower, onemklTranspose trans,
                 int64_t n, int64_t k, double _Complex alpha, const double _Complex *a,
                 int64_t lda, const double _Complex *b, int64_t ldb, double _Complex beta,
-                float _Complex *c, int64_t ldc);
+                double _Complex *c, int64_t ldc);
 
 // Supported Level-1: Nrm2
 void onemklDnrm2(syclQueue_t device_queue, int64_t n, const double *x,
